avoid kubelet remove too many pod at once

+++ pkg/kubelet/kubelet.go
@@ -361,6 +361,7 @@ func NewMainKubelet(kubeCfg *kubeletconfiginternal.KubeletConfiguration,
 	nodeLabels map[string]string,
 	seccompProfileRoot string,
 	bootstrapCheckpointPath string,
+	podNumberAvoidToRemoveAtOnce int32,
 	nodeStatusMaxImages int32) (*Kubelet, error) {
 	if rootDirectory == "" {
 		return nil, fmt.Errorf("invalid root directory %q", rootDirectory)
@@ -1227,6 +1228,9 @@ type Kubelet struct {
 
 	// Handles RuntimeClass objects for the Kubelet.
 	runtimeClassManager *runtimeclass.Manager
+
+	// Pod number avoid to remove at once, this is used by circuit-breaker when disrupation occured
+	podNumberAvoidToRemoveAtOnce int32
 }
 
 // setupDataDirs creates:
@@ -2106,6 +2110,27 @@ func (kl *Kubelet) HandlePodUpdates(pods []*v1.Pod) {
 // being removed from a config source.
 func (kl *Kubelet) HandlePodRemoves(pods []*v1.Pod) {
 	start := kl.clock.Now()
+
+	podWithoutDeleteTimestampCount := 0
+	for _, pod := range pods {
+		if pod.DeletionTimestamp == nil {
+			podWithoutDeleteTimestampCount++
+		}
+	}
+
+	// if there are cluster-level disruption, such as apiserver is misconfigured and misconnected to an wrong ETCD,
+	// the apiserver would like to send wrong control command and delete all pods belonging to this node.
+	// As last defense of running pod, kubelet should have some sanity check to avoid execute a wrong command.
+	// So when kubelet wants to remove a lots of pods which do not have deleteTimestamp assigned,
+	// we should stop the process, trigger the circuit breaker mechanism, and alert the maintainer to check
+	// what happened
+	if podWithoutDeleteTimestampCount > int(kl.podNumberAvoidToRemoveAtOnce) {
+		metrics.RemoveTooManyPodAtOnce.Inc()
+		klog.Errorf("Circuit breaker has been triggered to avoid remove more than %d pods at once, "+
+			"list of pods wanted to be removed: %s", kl.podNumberAvoidToRemoveAtOnce, format.Pods(pods))
+		return
+	}
+
 	for _, pod := range pods {
 		kl.podManager.DeletePod(pod)
 		if kubetypes.IsMirrorPod(pod) {
